Make it possible to use a header or a query parameter
Previously the SwitchUserStatelessBundle would only consider request
headers. This commit adds the ability to use a query parameter as well,
so the feature can be used in contexts where it's not possible to add
custom headers to the query.

To do so we introcude the new 'query_parameter' setting, which is empty
by default to keep the old behaviour. The old 'parameter' parameter is
renamed to 'header' to avoid confusion. This naming is somewhat
consistent with the naming used in LexikJwtAuthenticationBundle.

If the header or query parameter settings are set to an empty value,
then the related source is not considered.